# Stage_1_Insta-Analyzer-master-Sta/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Instagram Analyzer Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BotStatus {
  running: boolean;
  botName?: string;
  botUsername?: string;
  /** Uptime in seconds */
  uptime?: number;
  totalAnalyses: number;
}

export interface BotActionResult {
  success: boolean;
  message: string;
  running: boolean;
}

export interface AnalyzeRequest {
  username: string;
}

export interface AnalysisResult {
  username: string;
  fullName?: string;
  bio?: string;
  followers: number;
  following: number;
  postsCount: number;
  isVerified: boolean;
  isPrivate: boolean;
  isBusinessAccount: boolean;
  profilePicUrl?: string;
  engagementRate: number;
  avgLikes?: number;
  avgComments?: number;
  accountCategory?: string;
  externalUrl?: string;
  analyzedAt: string;
}

export interface AnalysisRecord {
  id: number;
  username: string;
  fullName?: string;
  followers: number;
  following: number;
  postsCount: number;
  isVerified: boolean;
  isPrivate: boolean;
  isBusinessAccount: boolean;
  engagementRate: number;
  avgLikes?: number;
  avgComments?: number;
  accountCategory?: string;
  profilePicUrl?: string;
  /** Telegram username or chat ID that requested this analysis */
  requestedBy?: string;
  analyzedAt: string;
}

export type BotStatsTopProfilesItem = {
  username: string;
  count: number;
};

export interface BotStats {
  totalAnalyses: number;
  uniqueUsers: number;
  uniqueProfiles: number;
  todayAnalyses: number;
  topProfiles: BotStatsTopProfilesItem[];
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetRecentAnalysesParams = {
  limit?: number;
};