# Stage_1_Insta-Analyzer-master-Sta/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Instagram Analyzer Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/recent:
    get:
      operationId: getRecentAnalyses
      tags: [analytics]
      summary: Get recent analyses
      description: Returns the most recent Instagram profile analyses
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of recent analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnalysisRecord"
  /analytics/analyze:
    post:
      operationId: analyzeProfile
      tags: [analytics]
      summary: Analyze an Instagram profile
      description: Fetches and analyzes an Instagram profile by username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeRequest"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    AnalyzeRequest:
      type: object
      properties:
        username:
          type: string
      required:
        - username
    AnalysisResult:
      type: object
      properties:
        username:
          type: string
        fullName:
          type: string
        bio:
          type: string
        followers:
          type: integer
        following:
          type: integer
        postsCount:
          type: integer
        isVerified:
          type: boolean
        isPrivate:
          type: boolean
        isBusinessAccount:
          type: boolean
        profilePicUrl:
          type: string
        engagementRate:
          type: number
        avgLikes:
          type: number
        avgComments:
          type: number
        accountCategory:
          type: string
        externalUrl:
          type: string
        analyzedAt:
          type: string
      required:
        - username
        - followers
        - following
        - postsCount
        - isVerified
        - isPrivate
        - isBusinessAccount
        - engagementRate
        - analyzedAt
    AnalysisRecord:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        followers:
          type: integer
        following:
          type: integer
        postsCount:
          type: integer
        isVerified:
          type: boolean
        isPrivate:
          type: boolean
        isBusinessAccount:
          type: boolean
        engagementRate:
          type: number
        avgLikes:
          type: number
        avgComments:
          type: number
        accountCategory:
          type: string
        profilePicUrl:
          type: string
        requestedBy:
          type: string
          description: Telegram username or chat ID that requested this analysis
        analyzedAt:
          type: string
      required:
        - id
        - username
        - followers
        - following
        - postsCount
        - isVerified
        - isPrivate
        - isBusinessAccount
        - engagementRate
        - analyzedAt
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message