# Stage_1_Insta-Analyzer-master-Sta/lib/api-zod/src/generated/types/analysisRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Instagram Analyzer Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisRecord {
  id: number;
  username: string;
  fullName?: string;
  followers: number;
  following: number;
  postsCount: number;
  isVerified: boolean;
  isPrivate: boolean;
  isBusinessAccount: boolean;
  engagementRate: number;
  avgLikes?: number;
  avgComments?: number;
  accountCategory?: string;
  profilePicUrl?: string;
  /** Telegram username or chat ID that requested this analysis */
  requestedBy?: string;
  analyzedAt: string;
}