# Stage_1_Insta-Analyzer-master-Sta/lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Instagram Analyzer Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisResult {
  username: string;
  fullName?: string;
  bio?: string;
  followers: number;
  following: number;
  postsCount: number;
  isVerified: boolean;
  isPrivate: boolean;
  isBusinessAccount: boolean;
  profilePicUrl?: string;
  engagementRate: number;
  avgLikes?: number;
  avgComments?: number;
  accountCategory?: string;
  externalUrl?: string;
  analyzedAt: string;
}